Implementing inference
